backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
requests==2.31.0
python-dotenv==1.0.0
aiohttp==3.9.1
python-multipart==0.0.6
emails==0.6
twilio==8.10.0
sqlalchemy==2.0.23
databases==0.8.0
aiosqlite==0.19.0